nodes: add IOVariable node
Currently unused, this node will be used to represent data passing
between program stages. This is removed from the responsibility of the
user as the syntax for that inter-stage communication is backend
specific.

While defining macros such as ngl_in/ngl_out to remap to in/out/varying
could do the trick for the differences between GL/GLES flavours, it
wouldn't do for Vulkan where specifying location is mandatory.

An alternative could have been to make ngl_in/ngl_out macros take a
mandatory location, but this was considered:
- more error prone (location increment is not always obvious)
- less future-proof (different syntaxes may arise)
- more risky (invalid location would work most of the time if tested
  with a non-modern backend because they would be ignored)
- less consistent with how other inputs/outputs are handled (uniforms,
  blocks, textures, etc)